finished the map function of our rocket
